XWIKI-19711: Webjar urls can't be resolved without calling the script service
