+++ xwiki-platform-core/xwiki-platform-webjars/xwiki-platform-webjars-api/src/main/resources/META-INF/components.txt
@@ -5,3 +5,4 @@ org.xwiki.webjars.internal.FilesystemResourceReferenceSerializer
 org.xwiki.webjars.script.WebJarsScriptService
 org.xwiki.webjars.internal.filter.LessWebJarsResourceFilter
 org.xwiki.webjars.internal.filter.VelocityWebJarsResourceFilter
+org.xwiki.webjars.internal.DefaultWebJarsUrlResolver
